Fix left-stick click to always show ability info (#57)
This overrides the base-game behaviour for left-stick click, which
simply checks whether the selected ability is at a higher rank than the
soldier's current rank. The new behaviour is to open the ability info
panel unless the ability is hidden.

This brings the controller behaviour into line with the mouse-and-
keyboard behaviour.

Fixes #57.